fix: loading the socket script in delayed phase
