+++ scripts/delayed.js
@@ -4,6 +4,8 @@ import {
   // eslint-disable-next-line import/named
   defaultAnalyticsLoadDisabled,
   loadAdobeLaunchAndGTM,
+  // eslint-disable-next-line import/named
+  loadStockFeed,
 } from './blocks-utils.js';
 
 // Core Web Vitals RUM collection
@@ -43,3 +45,5 @@ if (!isSidekickLibrary) {
   loadScript('/scripts/mockxmlhttprequest.js', { type: 'module' });
   loadScript('https://icici-securities.allincall.in/files/deploy/embed_chatbot_11.js?version=1.1');
 }
+
+loadStockFeed();
